requirements.txt
praw==7.7.1
pyvirtualdisplay==2.2
pytube==10.9.2
google-api-python-client==2.24.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
flask==2.0.2
elevenlabs==0.2.24
pydub==0.25.1
simpleaudio==1.0.4
moviepy==1.0.3
Pillow==10.0.0
